requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.21
mysqlclient==2.2.0
PyJWT==2.8.0